separation classes
